chore: consolidate services into single property of main class
