+++ src/services/index.ts
@@ -1,3 +1,3 @@
 export * from './search-service'
-export * from './generic-service'
+export * from './retrieval-service'
 export * from './resolution-service'
